Move MutexLock class into own file
